travis: Build on ARM64, IBM Power and IBM Z architectures
IBM Z is big-endian, IBM Power runs in little-endian mode.

Botan has issues with GCC (target pragma ‘*’ is invalid) on ARM64 and
IBM Power (also seems to ignore CC to select the compiler), while wolfSSL
can't be compiled successfully on IBM Z.

libunwind is not available for x390x, but since we explicitly disable
such backtraces it's not necessary anyway.